refactor: alterando a estrutura de pastas
